RDP 5.0 bulk compressor use Hash table manager.

+++ tests/core/RDP/test_mppc.cpp
@@ -117,9 +117,10 @@ BOOST_AUTO_TEST_CASE(TestMPPC_enc)
 
 BOOST_AUTO_TEST_CASE(TestBitsSerializer)
 {
-    uint8_t outputBuffer[256] ={};
-    int bits_left = 8;
-    int opb_index = 0;
+    uint8_t  outputBuffer[256] ={};
+    uint8_t  bits_left = 8;
+    uint16_t opb_index = 0;
+
     insert_n_bits_40_50(2, 3, outputBuffer, bits_left, opb_index);
     BOOST_CHECK_EQUAL(6, bits_left);
     BOOST_CHECK_EQUAL(0, opb_index);
